refactor: isolate Chrome host smoke
